services/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.2
openpyxl==3.1.2
pyarrow==15.0.2
scikit-learn==1.4.2
pydantic==1.10.13
python-multipart==0.0.9
scipy
numpy
joblib==1.4.2
boto3
python-dotenv